refactor(core): refine wasm preserve handling
