+++ packages/core/src/wasm/path_utils.ts
